+++ ui/Modal/Org/ConfigureEns/UpdateMetadata.svelte
@@ -12,7 +12,7 @@
 
   import { Modal, TextInput, Tooltip } from "ui/DesignSystem";
 
-  import ButtonRow from "./shared/ButtonRow.svelte";
+  import ButtonRow from "./ButtonRow.svelte";
 
   export let onSubmit: () => void;
   export let registration: ensResolver.Registration;
